Move cryotank_index into cryotank
